delete client manager, rename interfaces.py
