+++ src/compas_fab/backends/__init__.py
@@ -66,7 +66,6 @@ Exceptions
 
 """
 
-from .client_manager import *           # noqa: F401,F403
 from .exceptions import *               # noqa: F401,F403
 from .tasks import *                    # noqa: F401,F403
 from .ros.backend_features import *     # noqa: F401,F403
