+++ src/compas_fab/backends/ros/backend_features/move_it_add_attached_collision_mesh.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from compas_fab.backends.backend_feature_interfaces import AddAttachedCollisionMesh
+from compas_fab.backends.interfaces import AddAttachedCollisionMesh
 from compas_fab.backends.ros.messages import AttachedCollisionObject
 from compas_fab.backends.ros.messages import CollisionObject
