+++ src/compas_fab/backends/ros/backend_features/move_it_planning_scene.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 
 from compas.utilities import await_callback
 
-from compas_fab.backends.backend_feature_interfaces import GetPlanningScene
+from compas_fab.backends.interfaces import GetPlanningScene
 from compas_fab.backends.ros.messages import GetPlanningSceneRequest
 from compas_fab.backends.ros.messages import GetPlanningSceneResponse
 from compas_fab.backends.ros.messages import PlanningSceneComponents
