+++ src/compas_fab/backends/ros/backend_features/move_it_remove_collision_mesh.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from compas_fab.backends.backend_feature_interfaces import RemoveCollisionMesh
+from compas_fab.backends.interfaces import RemoveCollisionMesh
 from compas_fab.backends.ros.messages import CollisionObject
 
 __all__ = [
