+++ src/compas_fab/backends/vrep/backend_features/vrep_plan_motion.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 
 from timeit import default_timer as timer
 
-from compas_fab.backends.backend_feature_interfaces import PlanMotion
+from compas_fab.backends.interfaces import PlanMotion
 from compas_fab.backends.vrep.helpers import config_from_vrep
 from compas_fab.backends.vrep.helpers import config_to_vrep
 from compas_fab.backends.vrep.helpers import floats_to_vrep
