+++ src/compas_fab/backends/vrep/backend_features/vrep_remove_collision_mesh.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 
 from compas_fab.backends.vrep.helpers import DEFAULT_OP_MODE
 from compas_fab.backends.vrep.remote_api import vrep
-from compas_fab.backends.backend_feature_interfaces import RemoveCollisionMesh
+from compas_fab.backends.interfaces import RemoveCollisionMesh
 
 
 __all__ = [
